refactor(site/src): fold chat actions into a ChatMoreActions dropdown
The chat actions menu was rendered into two different menu primitives, a
dropdown and a row right-click context menu, so the items lived in a
ChatActionsMenuItems component that took the Item and Separator components
as props and left each call site to build its own dropdown.

Right-clicking a row also suppressed the browser's native context menu on
the row's NavLink, which removed "Open link in new tab" for every chat.

Remove right-click access and give the component the dropdown it names:
ChatMoreActions owns the trigger, the menu and the items, matching
WorkspaceMoreActions and AgentDevcontainerMoreActions. The Item and
Separator injection is no longer needed.

The two ChatsSidebar stories named after the context menu already drove the
row's ellipsis dropdown, so they are renamed rather than rewritten.
